Finishing storing and verifying payment.

+++ website/charity/models/payment.py
@@ -7,9 +7,6 @@ from charity.models.charity_profile import CharityProfile
 class Payment(models.Model):
     user_profile = models.ForeignKey(UserProfile, related_name='payments', null=True)
     charity_profile = models.ForeignKey(CharityProfile, related_name='payments')
-    gross = models.IntegerField()
+    gross = models.FloatField()
     currency = models.CharField(max_length=10, null=True)
     txn_id = models.CharField(max_length=100, null=False)
-
-
-
